requirements.txt
streamlit
streamlit-folium
fastapi
uvicorn
osmnx
networkx
folium
requests
python-dotenv
